fixed some working on calc_pi.
Updated calc_private to be compatible with snpR 1.0. Should also be faster, uses some reshape tools to do things quickly and without any loops.

Updated calc_ho to be compatible with snpR 1.0, and should be faster as well.

Most of the basic stats are done, need to do fst, taijima's D.