Removed stickFORMATs and references to it--too big to be worth it.
Added a header_cols argument to import.snpR.data and read_delimited_snps to support automatically grabbing some columns as snp metadata header columns when reading in a file.

Fixed a test in test-subset that didn't work due to scope issues. Since the fix uses <<-, which is very not ideal, the test is skipped on cran.